Reorganise overworlds to set up for expansion
The following still needs to be changed in order to use more than 256 overworlds:

// global.fieldmap.h
struct EventObjectTemplate
struct EventObject
// asm/macros/map.inc
.macro object_event

However doing so at present causes the game to crash, and requires all code within /asm that references gSaveBlock1 to be decompiled.

+++ include/berry.h
@@ -23,7 +23,7 @@ enum
 
 #define NUM_BERRIES 44
 
-extern const u8 *const gBerryTreeEventObjectGraphicsIdTablePointers[];
+extern const u16 *const gBerryTreeEventObjectGraphicsIdTablePointers[];
 extern const struct SpriteFrameImage *const gBerryTreePicTablePointers[];
 extern const u8 *const gBerryTreePaletteSlotTablePointers[];
